requirements.txt
streamlit
playwright
pandas
pyyaml
rapidfuzz
pillow
